+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -16,10 +16,10 @@ import {
   UrlManager,
   buildSearchEngine,
   SearchEngine,
-  loadSearchConfigurationActions,
   SearchEngineConfiguration,
   SearchStatus,
   buildSearchStatus,
+  loadSearchConfigurationActions,
 } from '@coveo/headless';
 import {Bindings, InitializeEvent} from '../../utils/initialization-utils';
 import i18next, {i18n} from 'i18next';